feat: configure test environment variables in application-test.properties
This commit introduces a new file, `application-test.properties`, to
manage environment variables specifically for the test environment.
This allows us to configure the test environment separately from the
development and production environments.

The `deploy.yaml` workflow is updated to activate the `test` Spring
profile during the test job. The environment variables are removed
from the workflow file and configured in the
`application-test.properties` file.